--- tests/new_test.py
@@ -1,84 +1,0 @@
-import pytest
-import os
-from pykern.pkdebug import pkdp
-from sirepo import simulation_db
-import sirepo.uri
-
-
-def test_favicon(fc):
-    from pykern.pkcollections import PKDict
-    from pykern.pkunit import pkeq
-    import sirepo.resource
-    import sys
-
-    for a, p in PKDict(favicon="favicon.ico", faviconPng="favicon.png").items():
-        r = fc.sr_get(a)
-        with open(sirepo.resource.static("img", p), "rb") as f:
-            s = f.read()
-            pkeq(len(s), len(r.data))
-
-
-def test_find_by_name(fc):
-    from pykern.pkcollections import PKDict
-    from pykern.pkunit import pkeq
-
-    fc.sr_get(
-        "findByName",
-        PKDict(
-            simulation_type="srw",
-            application_mode="default",
-            simulation_name="Undulator Radiation",
-        ),
-        redirect=False,
-    ).assert_http_redirect("/srw#/findByName/default/Undulator%20Radiation")
-
-
-def test_custom_errors(fc):
-    from pykern.pkunit import pkre
-
-    for k, v in simulation_db.SCHEMA_COMMON.customErrors.items():
-        r = fc.sr_get(sirepo.uri.local_route(fc.sr_sim_type, v.route))
-        r.assert_success()
-        with open(sirepo.resource.static("html", v.url), "r") as f:
-            s = f.read()
-            pkre(v.msg, s)
-
-
-def test_simulation_redirect(fc):
-    fc.sr_get("simulationRedirect", redirect=False).assert_http_redirect("simulations")
-
-
-def test_error_logging(fc):
-    from pykern.pkcollections import PKDict
-    from pykern.pkunit import pkeq
-
-    r = fc.sr_post("errorLogging", PKDict(message="error message"))
-    pkeq("ok", r.state)
-
-
-def test_simulation_schema(fc):
-    from pykern.pkcollections import PKDict
-    from pykern.pkunit import pkexcept
-
-    r = fc.sr_post_form(
-        "simulationSchema",
-        data=PKDict(simulationType=fc.sr_sim_type),
-    )
-
-    for k in ("model", "view"):
-        assert k in r.keys()
-
-    with pkexcept("unexpected status"):
-        fc.sr_post_form(
-            "simulationSchema",
-            data=PKDict(simulationType="xyzzy"),
-        )
-
-
-def test_srw_light(fc):
-    from pykern.pkunit import pkre
-
-    r = fc.sr_get("srwLight")
-    r.assert_success()
-    pkre("<!DOCTYPE html>", r.data)
-    pkre('data-ng-app="SRWLightGateway"', r.data)
